Separate roles and permissions, moving permissions into the core spec
Permissions serve a vital role in the communication between the client
and server: they let the client know what actions it can take before
taking them. For this reason, clients would be acting blind without
permissions in the core specification.

While server implementations might often choose to apply permission
based on roles, this information is irrelevant to the client - the
client needs to know what *it* can do, not what other users can do.

With permissions gone, roles become a rather odd object. Factions
already serve the role of grouping users which is fundamentally all
roles do now. In practice, the two are used differently but it might be
worth changing roles significantly in the near future.