Replace width and height with shape
Shape allows for much more flexible and intersting board shapes while
being easy to use as a standard width + height once it is understood.

For clarity, clients only supporting a width + height arrangement should
check that the shape is of length 1 and that its one item is of size 2.
Servers only supporting this should likewise only indicate such a shape.