Rewrite authentication extension
My knowledge on openid and identity servers in general has improved a
lot recently, so I feel in a much better position to understand what
needs to be specified in this extensions. I didn't really know what I
was doing before, so I suspect it has gone from making no sense at all
to being understandable — even if it is a bit vague.