Require that position is specified directly rather than as coordinates
This change was spurred from it making the serverside code a bit neater
in practice. My hope is that it will have minimal impact on client code,
since working in raw positions is probably as easy on both sides.

It also has the added benefit of removing the somewhat strange sequence
of arbitrary path separators, at the cost of being harder for humans to
read.