Remove permissions-changed event
This may need some sort of replacement later, but for now it seems out
of place among other board events given it's more site-wide.

Those paying attention to the various extensions may notice a general
neglect for them. The plan is to update them as they are implemented in
a reference server that is currenly being worked on. This reference
server is guiding my thoughts on how this API should work.