Remove idle user count from users endpoint
Idle users makes sense if we're talking about websocket connections,
which Pxls original uses to keep track of this.

After thinking about this, it seems like it makes more sense to track
"active" users as not the number of user-unique websocket connections,
but as the number of user-unique placements within a timeframe. This
obviously gives no way to measure idle users in a worthwhile sense, so
the field wouldn't make sense for this case.

Since the number if idle users means little anyway, this just removes it
altogether.