Separate additional board buffers into extensions
Since `colors` is the only technically required buffer, making the
others optional allows simpler clients and servers which need not resort
to workarounds to skip implementing other buffers.

Since clients can no longer be relied on to know about the initial board
buffer, the `board-update` event needs to specify which parts of the
color buffer changed. For large initial changes, server implementations
may prefer not to send this change over the socket. To facilitate this,
the `Change` object can now simply specify a range to invalidate rather
than the set of new values.