Separate some extensions functionality, re-assess which extensions to include (#8)
* Remove chat extensions

Chatrooms are widely implemented. I see little compelling reason to add
a new sub-par standard for chatrooms. Clients may desire to implement
their own chat functionality, but the server (and thus the networking
between the two) need not know about it.

* Separate additional board buffers into extensions

Since `colors` is the only technically required buffer, making the
others optional allows simpler clients and servers which need not resort
to workarounds to skip implementing other buffers.

Since clients can no longer be relied on to know about the initial board
buffer, the `board-update` event needs to specify which parts of the
color buffer changed. For large initial changes, server implementations
may prefer not to send this change over the socket. To facilitate this,
the `Change` object can now simply specify a range to invalidate rather
than the set of new values.

* Separate user count into a separate extension

Since knowing the current user count is entirely optional to the
function of the service, it makes most sense to be an extension rather
than part of the core specification.

* Separate board updating functions into separate extension

To allow for simpler server implementations which do not wish to support
the complexity of dynamic board management, this functionality can be
made into an extension.

* Add missing info response for user_count

* Make values and length mutually exclusive in board updates

The previous specification encoded redundant data and wasn't as clear.

* Specify pixel access with position rather than coordinates

This is a consitency change across extensions as it was already the case
for the core protocol.

* Rename announcements to notices

This more accurately conveys the use and purpose of the extensions.

* Fix typo

* Improve clarity of phrasing in roles extension

* Improve phrasing and naming of faction joining

* Allow ban expiry and reason to be explicity null

* Add details for permissions handling with websocket connections

* Use custom close codes for websocket errors

This helps simplify things

* Add missing full-stops

* Update websocket endpoint from /ws to /socket

This is more consistent, but the name socket is not quite as descriptive
as I'd like. This may change to something like "events" or "listener".

* Update user extension to use references

Also fixes a couple of mistakes:
- Since placements are extended, the request change to pixels is
  redundant.
- The user update endpoint was defined as a partial user. This might
  have been inentional at the time, but it seems incorrect to allow
  editing of the `created_at` field.

* Capitalize ID in authentication

* Update older extensions to use references and remove IDs

This brings outdated definitions in line with the newer specification
base and extensions.

* Move some faction customization into a new extension

Still not entirely sure about the exact fields here: there's definitely
room to add others or re-evaluate the use of the existing ones.

* Use error table for current user endpoint

* Improve wording in core spec

* Cover /socket and board sockets in authentication

* Update errors for user patch

* Improve consistency of responses and errors for roles

* Improve wording regarding using object addresses

* Send User Reference instead of view for update packet