Change default boards from being a redirect to a reference
In the infinite wisdom of the web standards group, it is impossible —
even with server cooperation — for a browser environment to get the
location of a redirect. This means that all requests using the default
board feature would need to actually follow the redirect ever time
unless cached, which is wasteful.

The upside is that this made me think a little about this and it is
probably more sensible to have it return a reference since then if the
default board changes, the client's requests don't suddenly reference a
different board to the one it was tracking interally.